Make the r3 node embedding-dependent
Update schnyder_realizer.html and its images accordingly

Previously the r1, r2 and r3 node adjacency lists in the
embedding were, respectively, the 1st, the 2nd and the 3rd ones
the user has to insert.
Now the user is required to insert the r3 adjacency list as the
last one.